requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.0
facenet-pytorch==2.5.3
retina-face==0.0.13
tensorflow
pinecone
python-dotenv==1.0.1